Move page not found to newly created pages folder.
